Made 2-hop-test-50percent-drop actually drop packets.
