+++ src/test/cooja/2-hop-test-50percent-drop/2hop-test-50percent-drop.csc
@@ -14,8 +14,8 @@
       se.sics.cooja.radiomediums.UDGM
       <transmitting_range>55.0</transmitting_range>
       <interference_range>100.0</interference_range>
-      <success_ratio_tx>1.0</success_ratio_tx>
-      <success_ratio_rx>1.0</success_ratio_rx>
+      <success_ratio_tx>0.6</success_ratio_tx>
+      <success_ratio_rx>0.6</success_ratio_rx>
     </radiomedium>
     <events>
       <logoutput>40000</logoutput>
