+++ src/test/cooja/2-hop-test-50percent-drop/Makefile
@@ -4,17 +4,23 @@ CONTIKI = ../../../../lib/contiki
 K = 2
 MAX_NODES = 20
 MAX_EDGES = 100
+DISCOVERY_INTERVAL_MIN = 2
+DISCOVERY_INTERVAL_MAX = 4
 
 CFLAGS += -Wall
 
 CFLAGS += -DK=$(K)
 CFLAGS += -DMAX_NODES=$(MAX_NODES)
 CFLAGS += -DMAX_EDGES=$(MAX_EDGES)
+CFLAGS += -DDISCOVERY_INTERVAL_MIN=$(DISCOVERY_INTERVAL_MIN)
+CFLAGS += -DDISCOVERY_INTERVAL_MAX=$(DISCOVERY_INTERVAL_MAX)
+
 
 PROJECT_SOURCEFILES+=../../../graph.c
 PROJECT_SOURCEFILES+=../../../graph-operations.c
 PROJECT_SOURCEFILES+=../../../serialize.c
 PROJECT_SOURCEFILES+=../../../pbroadcast.c
+PROJECT_SOURCEFILES+=../../../neighbor-discovery.c
 CFLAGS += -I../../../
 
 CFLAGS += -D__K_HOP_DEBUG__=1
